requirements.txt
beautifulsoup4
black
coverage
flake8
flask
flask_login
flask_testing
flask_wtf
fpdf
isort
numpy
pandas
pymongo
pytest
python-dotenv
scikit-learn
selenium
webdriver-manager
werkzeug
wtforms
flask_login
bson